+++ src/front/styles/register.scss
@@ -1,0 +1,188 @@
+@import "~bootstrap/scss/bootstrap.scss";
+
+body {
+    margin: 0;
+    color: rgba(107, 101, 103, 0.781);
+    background:  rgba(245, 222, 230, 0.76);
+    font: 600 1rem/1.125 'Open Sans', sans-serif
+}
+
+.login-box {
+    width: 100%;
+    margin: auto;
+    max-width: 100rem;
+    min-height: 90rem;
+    position: relative;
+    background: rgb(207, 174, 205);
+    box-shadow: 0 0.75rem 0.938rem 0 rgba(27, 26, 26, 0.24), 0 17px 50px 0 rgba(0, 0, 0, .19)
+}
+
+.login-snip {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    padding: 5.625rem 4.375rem 3.125rem 4.357rem;
+    background: rgb(219, 189, 189)
+}
+
+.login-snip .sign-in,
+.login-snip .sign-up,
+.login-space .group .check {
+    display: none
+}
+
+.login-snip .tab,
+.login-space .group .label,
+.login-space .group .button {
+    text-transform: uppercase
+}
+
+.login-snip .tab {
+    font-size: 1.375rem;
+    margin-right: 0.938rem;
+    padding-bottom: 0.313rem;
+    margin: 0 0.938rem 0.625rem 0;
+    display: inline-block;
+    border-bottom: 0.125rem solid transparent
+}
+
+.login-snip .sign-in:checked+.tab,
+.login-snip .sign-up:checked+.tab {
+    color: #fff;
+    border-color: #7b7c7e
+}
+
+.login-space {
+    min-height: 21.563rem;
+    position: relative;
+    perspective: 62.5;
+    transform-style: preserve-3d
+}
+
+.login-space .group {
+    margin-bottom: 0.938rem;
+    border-radius: 1.563rem
+}
+
+.login-space .group .label,
+.login-space .group .input,
+.login-space .group .button {
+    width: 100%;
+    color: rgb(99, 96, 96);
+    display: block
+}
+
+.login-space .group .input,
+.login-space .group .button {
+    border: none;
+    padding: 0.938rem 1.25rem;
+    border-radius: 1.563rem;
+    background: rgba(255, 255, 255, 0.863)
+}
+
+
+.login-space .group .label {
+    color: rgb(66, 60, 60);
+    font-size: 0.75rem
+}
+
+.login-space .group .button {
+    background: #919294
+}
+
+.login-space .group label .icon {
+    width: 1rem;
+    height: 1rem;
+    border-radius: 0.125rem;
+    position: relative;
+    display: inline-block;
+    background: rgba(255, 255, 255, .1)
+}
+
+.login-space .group label .icon:before,
+.login-space .group label .icon:after {
+    content: '';
+    width: 0.625rem;
+    height: 0.125rem;
+    background: rgba(255, 255, 255, 0.288);
+    position: absolute;
+    transition: all .2s ease-in-out 0s
+}
+
+.login-space .group label .icon:before {
+    left: 0.188rem;
+    width: 0.313rem;
+    bottom: 0.375rem;
+    transform: scale(0) rotate(0)
+}
+
+.login-space .group label .icon:after {
+    top: 0.375rem;
+    right: 0;
+    transform: scale(0) rotate(0)
+}
+
+.login-space .group .check:checked+label {
+    color: rgba(255, 255, 255, 0.342)
+}
+
+.login-space .group .check:checked+label .icon {
+    background: #1161ee
+}
+
+.login-space .group .check:checked+label .icon:before {
+    transform: scale(1) rotate(45deg)
+}
+
+.login-space .group .check:checked+label .icon:after {
+    transform: scale(1) rotate(-45deg)
+}
+
+.login-snip .sign-in:checked+.tab+.sign-up+.tab+.login-space .login {
+    transform: rotate(0)
+}
+
+.login-snip .sign-up:checked+.tab+.login-space .sign-up-form {
+    transform: rotate(0)
+}
+
+*,
+:after,
+:before {
+    box-sizing: border-box
+}
+
+.clearfix:after,
+.clearfix:before {
+    content: '';
+    display: table
+}
+
+.clearfix:after {
+    clear: both;
+    display: block
+}
+
+a {
+    color: inherit;
+    text-decoration: none
+}
+
+.hr {
+    height: 0.125rem;
+    margin: 3.75rem 0 10rem 0;
+    background: rgba(255, 255, 255, .2)
+}
+
+.foot {
+    text-align: center
+}
+
+.card {
+    width: 31.25rem;
+    left: 6.25rem
+}
+
+::placeholder {
+    color: #b3b3b3
+}
